requirements.txt
pytest==7.2.2
pytest-cov==4.1.0
pytest-django==4.5.2; python_version < '3.8'
pytest-django==4.7.0; python_version >= '3.8'
pytest-xdist==3.5.0
